backend/pyproject.toml
[tool.poetry]
name = "escape-room-booking-backend"
version = "0.1.0"
description = "Booking system made with Django"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.*"
django-debug-toolbar = "^4.1.0"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.3"
django-cors-headers = "^4.3.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"